Update the build specification, CONTRIBUTING and the README for release v1.0.0.1 (#6)
### What does this Pull Request accomplish?

This PR contains changes on the build specification of `G Plug-in SDK
Generator` vi package that excludes the G Plug-in migrator tool, updates
the CONTRIBUTING.md with the repo information and removes the
information about G Plug-in migrator in the README.md file. Resets the
build version to 1.0.0.1 for creating a new release. Updates a library
name for maintaining consistency in G Plug-In generator.

### Why should this Pull Request be merged?

This PR should be merged to do a release with the G Plug-in generator
without G Plug-in migrator.

### What testing has been done?

Manual testing is done with the VIPM packages.

---------

Co-authored-by: AishwaryaLakshmiSS <aishwarya.lakshmi.solitontech.com>